Add new "length" attribute
It's now possible to describe the length of a variable sized field with
an arithmetic expression that may also include other field names.

An example would be the options field in an IPv4 packet

    #[packet]
    pub struct Ipv4 {
        ...
        header_length: u4,
        ...
        #[length = "header_length - 4"]
        options: Vec<Ipv4Option>,
        ...
    }

Within such a length expression only field names, integers and basic
arithmetic operators like + - * / % are allowed.