Add function to determine the packet_size from a struct
There's a new `packet_size` function that is implemented for the
packets in case the size of a packet can be determined from the
struct. This is the case when the sizes of all variable sized fields
are defined by a length expression (as opposed to a length function).

This makes it possible to create a new packet from a struct without
knowing its size up front. Example:

    #[packet]
    pub struct Simple {
        sz: u8,
        #[length = "sz"]
        #[payload]
        payload: Vec<u8>
    }

    let payload = vec![1, 2, 3, 4];
    let simple = Simple {
        sz: payload.len() as u8,
        payload: payload
    };

    let mut packet = vec![0; SimplePacket::packet_size(&simple)];
    MutableDcpOpenPacket::new(&mut packet[..]).unwrap().populate(simple);