compare on payload, not on full L2 data
We can't rely on L2 headers to be equal.

On OSX in lo0 ethetype was received as 0x0. I cannot argue it's incorrect to do
so, or incorrect not to do so. I can see why either approach may make sense,

and it's best if we don't rely on either.